requirements-tests.txt
-r requirements.txt
-e .

marshmallow-sqlalchemy  >= 0.8.0
marshmallow-peewee      >= 2.0.0
peewee                  >= 3.7.1
mongomock>=3.14.0
pymongo>=3.7.2

ipdb                    >= 0.11
pytest>=4.0.0
pytest-flask>=0.14.0
pytest-sugar>=0.9.2